+++ testBalloon-framework-core/src/commonTest/kotlin/de/infix/testBalloon/framework/core/TestConfigTests.kt
@@ -3,8 +3,10 @@ package de.infix.testBalloon.framework.core
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.currentCoroutineContext
 import kotlin.test.Test
+import kotlin.test.assertContains
 import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 
@@ -33,6 +35,36 @@ class TestConfigTests {
 
         TestConfig.invocation(TestInvocation.CONCURRENT).executeWrapped(testSuite) {
             assertEquals(TestInvocation.CONCURRENT, TestInvocation.current())
+        }
+    }
+
+    @Test
+    fun concurrencyAndTestScope() = withTestFramework {
+        val testSuite by testSuite("testSuite") {}
+
+        TestSession.DefaultConfiguration.executeWrapped(testSuite) {
+            assertEquals(TestInvocation.SEQUENTIAL, TestInvocation.current())
+            assertNotNull(TestScopeContext.current())
+        }
+
+        TestSession.DefaultConfiguration.invocation(TestInvocation.CONCURRENT).executeWrapped(testSuite) {
+            assertEquals(TestInvocation.CONCURRENT, TestInvocation.current())
+            assertNull(TestScopeContext.current())
+        }
+
+        assertFailsWith<IllegalArgumentException> {
+            TestConfig.invocation(TestInvocation.CONCURRENT).testScope(isEnabled = true).executeWrapped(testSuite) {}
+        }.also {
+            assertContains(it.message!!, "attempt was made to enable a 'TestScope' in combination with concurrent")
+        }
+
+        val testConfig = TestConfig.invocation(TestInvocation.CONCURRENT)
+            .addPermits(TestPermit.TEST_SCOPE_WITH_CONCURRENT_INVOCATION)
+            .testScope(isEnabled = true)
+        testConfig.parameterize(testSuite) // This is required to pick up permits
+        testConfig.executeWrapped(testSuite) {
+            assertEquals(TestInvocation.CONCURRENT, TestInvocation.current())
+            assertNotNull(TestScopeContext.current())
         }
     }